Move clean-up query functions to separate module
